+++ ShapeCrawler.IntegrationTests/UnitTest1.cs
@@ -1,0 +1,9 @@
+namespace ShapeCrawler.IntegrationTests;
+
+public class UnitTest1
+{
+    [Fact]
+    public void Test1()
+    {
+    }
+}
