+++ ShapeCrawler.sln
@@ -12,6 +12,8 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ShapeCrawler", "ShapeCrawle
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ShapeCrawler.UnitTests", "ShapeCrawler.UnitTests\ShapeCrawler.UnitTests.csproj", "{1D037182-15DF-4700-9E13-63CA8FED962D}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ShapeCrawler.IntegrationTests", "ShapeCrawler.IntegrationTests\ShapeCrawler.IntegrationTests.csproj", "{745E108F-2981-40DB-A607-ECE65C83F77D}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -31,6 +33,12 @@ Global
 		{1D037182-15DF-4700-9E13-63CA8FED962D}.Release|Any CPU.Build.0 = Release|Any CPU
 		{1D037182-15DF-4700-9E13-63CA8FED962D}.Test|Any CPU.ActiveCfg = Test|Any CPU
 		{1D037182-15DF-4700-9E13-63CA8FED962D}.Test|Any CPU.Build.0 = Test|Any CPU
+		{745E108F-2981-40DB-A607-ECE65C83F77D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{745E108F-2981-40DB-A607-ECE65C83F77D}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{745E108F-2981-40DB-A607-ECE65C83F77D}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{745E108F-2981-40DB-A607-ECE65C83F77D}.Release|Any CPU.Build.0 = Release|Any CPU
+		{745E108F-2981-40DB-A607-ECE65C83F77D}.Test|Any CPU.ActiveCfg = Debug|Any CPU
+		{745E108F-2981-40DB-A607-ECE65C83F77D}.Test|Any CPU.Build.0 = Debug|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
